refactor syncer code
