+++ crates/aranya-daemon/src/sync/mod.rs
@@ -1,0 +1,5 @@
+//! Aranya syncer for syncing Aranya graph commands.
+
+pub mod actions;
+pub mod task;
+pub mod tcp;
