+++ crates/aranya-daemon/tests/tests.rs
@@ -9,8 +9,8 @@
 
 use anyhow::{Context, Result};
 use aranya_daemon::{
-    aranya::Actions,
     policy::{Effect, Role},
+    sync::actions::Actions,
 };
 use serial_test::serial;
 use test_log::test;
